Split HTML secondary pane renderer tests
