+++ docs/DECISIONS.md
@@ -318,3 +318,4 @@
 - HTML tool-card renderer coverage lives in `WorkspaceHTMLToolCardRendererTests`, not the broad `WorkspaceSurfaceTests` catch-all. Tool-card output, approval actions, artifact chips, image/document/appshot previews, and transcript ordering should stay beside the focused renderer so artifact UI work does not inflate the generic workspace surface suite.
 - HTML terminal renderer coverage lives in `WorkspaceHTMLTerminalRendererTests`, not the broad `WorkspaceSurfaceTests` catch-all. Visible terminal pane markup, command rows, clear controls, and running/stopped status labels should stay beside terminal HTML so terminal UI work can evolve without expanding generic surface projection tests.
 - HTML review renderer coverage lives in `WorkspaceHTMLReviewRendererTests`, not the broad `WorkspaceSurfaceTests` catch-all. Static review-pane markup, review file/action/hunk/line/comment rendering, and review action data attributes should stay beside review HTML while review-surface data projection remains in focused workspace surface tests until it has a separate builder suite.
+- HTML secondary-pane renderer coverage lives in `WorkspaceHTMLSecondaryPaneRendererTests`, not the broad `WorkspaceSurfaceTests` catch-all. Extensions and Memories static HTML smoke tests should stay beside `WorkspaceHTMLSecondaryPaneRenderer`; browser HTML remains in browser-focused suites because the browser pane has a separate ownership boundary.
